requirements.txt
ipython==7.9.0
Pillow==6.2.1
